edit: changed message dates and added image sending

+++ src/client/client.ts
@@ -4,7 +4,10 @@ import Client from "../types/client.js";
 import Thread from "../types/thread.js";
 import { getInbox } from "../lib/inbox.js";
 import Message, { castMessage } from "../types/message.js";
-import sendMessage from "../lib/send-message.js";
+import sendMessage, {
+  sendImageFromUrl,
+  sendImageStoryFromUrl,
+} from "../lib/send-message.js";
 import {
   askCommand,
   askCredentials,
@@ -17,6 +20,7 @@ import getStories, { displayStoryData } from "../lib/stories.js";
 
 const commandDelete = async (thread: Thread, client: Client) => {
   const itemId = await askDeleteMessage(thread, client);
+  if (itemId == "-1") return;
   let instagramThread = client.ig.entity.directThread(thread.id);
   let deleteSpinner = ora("Deleting item").start();
   await instagramThread.deleteItem(itemId);
@@ -31,8 +35,8 @@ const commandStories = async (thread: Thread, client: Client) => {
   } else {
     let storiesSpinner = ora("Fetching stories").start();
     const stories = await getStories(client, thread.users[0].username);
-    storiesSpinner.stop()
-    console.log("Stories from " + thread.users[0].username)
+    storiesSpinner.stop();
+    console.log("Stories from " + thread.users[0].username);
     if (stories.length > 0) {
       stories.forEach((media, index) => {
         displayStoryData(media, index);
@@ -49,16 +53,30 @@ const handleMessage = async (
   thread: Thread
 ) => {
   if (messageText.length == 0) return false;
-  if (messageText == "/end" || messageText == "/exit" || messageText == "/stop") return true;
-  if (messageText == "/delete" || messageText == "/remove" || messageText == "/dlt" || messageText == "/rm") {
+  if (messageText == "/end" || messageText == "/exit" || messageText == "/stop")
+    return true;
+  if (
+    messageText == "/delete" ||
+    messageText == "/remove" ||
+    messageText == "/dlt" ||
+    messageText == "/rm"
+  ) {
     await commandDelete(thread, client);
     return false;
   }
-  if (messageText == "/story" || messageText == "/str" || messageText == "/sto") {
+  if (
+    messageText == "/story" ||
+    messageText == "/str" ||
+    messageText == "/sto"
+  ) {
     await commandStories(thread, client);
     return false;
   }
-  if (messageText == "/cmd" || messageText == "/command" || messageText == "/") {
+  if (
+    messageText == "/cmd" ||
+    messageText == "/command" ||
+    messageText == "/"
+  ) {
     let command = await askCommand();
     if (command == "exit") return true;
     if (command == "cancel") return false;
@@ -66,7 +84,31 @@ const handleMessage = async (
     if (command == "stories") await commandStories(thread, client);
     return false;
   }
-  const messageSpinner = ora("Sending message");
+  if (messageText.startsWith("/img") || messageText.startsWith("/image")) {
+    let imageUrl = messageText.split(" ")[1];
+    if (!imageUrl || imageUrl.length == 0) {
+      console.log("Invalid url");
+      return false;
+    }
+    await sendImageFromUrl(client, thread.id, imageUrl);
+    return false;
+  }
+  if (
+    messageText.startsWith("/simg") ||
+    messageText.startsWith("/storyimage")
+  ) {
+    let imageUrl = messageText.split(" ")[1];
+    if (!imageUrl || imageUrl.length == 0) {
+      console.log("Invalid url");
+      return false;
+    }
+    let imageStoryMessageSpinner = ora("Sending story image");
+    imageStoryMessageSpinner.start();
+    await sendImageStoryFromUrl(client, thread.id, imageUrl);
+    imageStoryMessageSpinner.stop();
+    return false;
+  }
+  let messageSpinner = ora("Sending message");
   messageSpinner.start();
   await sendMessage(client, thread.id, messageText);
   messageSpinner.stop();
@@ -75,16 +117,8 @@ const handleMessage = async (
 
 export const openThread = async (thread: Thread, client: Client) => {
   var conversationEnded = false;
-
   renderMessages(thread, client);
-  let threadSpinner = ora("Loading realtime").start();
-  await client.ig.realtime.connect({
-    irisData: await client.ig.feed.directInbox().request(),
-  });
-  threadSpinner.stop();
-
   client.ig.realtime.on("close", () => console.error("RealtimeClient closed"));
-
   client.ig.realtime.on("message", (messageRealtime: any) => {
     let messageData: any = messageRealtime.message;
     if (messageData.op == "add") {
@@ -102,7 +136,6 @@ export const openThread = async (thread: Thread, client: Client) => {
     let stop = await handleMessage(messageText, client, thread);
     conversationEnded = stop;
   }
-  client.ig.realtime.disconnect();
 };
 
 export const startInterface = async () => {
@@ -119,17 +152,40 @@ export const startInterface = async () => {
     client = await logClientSafe(credentials.username, credentials.password);
     clientSpinner.stop();
   }
+  let threadSpinner = ora("Loading realtime").start();
+  await client.ig.realtime.connect({
+    irisData: await client.ig.feed.directInbox().request(),
+  });
+
+  threadSpinner.stop();
+  const inboxSpinner = ora("Fetching threads").start();
+  var inbox: Thread[] = await getInbox(client);
+  inboxSpinner.stop();
+  // client.ig.realtime.on("message", (messageRealtime: any) => {
+  //   let messageData: any = messageRealtime.message;
+  //   if (messageData.op == "add") {
+  //     let message: Message = castMessage(messageData);
+  //     inbox.forEach(thread => {
+  //       if (thread.id == message.threadId)
+  //         thread.unread = true
+  //     })
+  //   }
+  // });
+  
   while (true) {
     console.clear();
-    const inboxSpinner = ora("Fetching threads").start();
-    const inbox: Thread[] = await getInbox(client);
-    inboxSpinner.stop();
     const threadIndex: number = await askThread(inbox);
     if (threadIndex == -1) break;
-    if (threadIndex == -2) continue;
+    if (threadIndex == -2) {
+      inboxSpinner.start();
+      inbox = await getInbox(client);
+      inboxSpinner.stop();
+      continue;
+    }
     const thread: Thread = inbox[threadIndex];
     await openThread(thread, client);
   }
+  client.ig.realtime.disconnect();
 };
 
 export default startInterface;
